requirements.txt
certifi==2019.9.11
charset-normalizer==2.0.12
click==8.0.4
Deprecated==1.2.13
fakeredis==1.7.1
idna==3.3
packaging==21.3
pipenv==2018.11.26
psutil==5.9.0
pydantic==1.9.0
pymongo==4.0.2
pyparsing==3.0.7
python-on-whales==0.37.0
PyYAML==6.0
redis==4.1.4
redislite==6.0.674960
requests==2.27.1
six==1.16.0
sortedcontainers==2.4.0
splunk-hec-handler==1.1.0
tqdm==4.63.0
typer==0.4.0
typing_extensions==4.1.1
urllib3==1.26.8
virtualenv==16.7.7
virtualenv-clone==0.5.3
wrapt==1.13.3
